requirements.txt
aiogram
apscheduler
python-dateutil
aiosqlite